functions/requirements.txt
firebase_functions~=0.1.0
firebase-admin==6.2.0
google-cloud-storage==2.14.0
python-dotenv==1.0.0